+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './usePagination';
 export * from './useResizeObserver';
 export * from './useElementSize';
+export * from './useWindowSize';
